+++ global_trotter_web/src/App.jsx
@@ -9,6 +9,7 @@ import Destinations from './pages/Destinations.jsx'
 import DestinationDetails from './pages/Destinationdetails.jsx'
 import Itineraries from './pages/Itineraries.jsx'
 import ItineraryDetails from './pages/itineraryDetails.jsx'
+import MapPage from './pages/MapPage.jsx'
 import Profile from './pages/Profile.jsx'
 import Favorites from './pages/Favorites.jsx'
 import MyDestinations from './pages/MyDestinations.jsx'
@@ -29,6 +30,7 @@ function App() {
           <Route path="/destinations/:id" element={<DestinationDetails />} />
           <Route path="/itineraries" element={<Itineraries />} />
           <Route path="/itineraries/:id" element={<ItineraryDetails />} />
+          <Route path="/map" element={<MapPage />} />
           <Route path="/profile" element={<Profile />} />
           <Route path="/favorites" element={<Favorites />} />
           <Route path="/my-destinations" element={<MyDestinations />} />